For this phase, you will keep a set of rocks and spawn new rocks into this set.  This requires the following steps:
Remove a_rock and replace it with rock_group. Initialize the rock group to an empty set. Modify your rock spawner to create a new rock (an instance of a Sprite object) and add it to rock_group.
Modify your rock spawner to limit the total number of rocks in the game at any one time. We suggest you limit it to 12. With too many rocks the game becomes less fun and the animation slows down significantly.
Create a helper function process_sprite_group. This function should take a set and a canvas and call the update and draw methods for each sprite in the group.
Call the process_sprite_group function on rock_group in the draw handler.